V1.6 - Interactive Menu!!!
- Added the "PluginMenuManager"!
+ When typing /sl or /sl _Menu, a menu will now open, featuring all of the plugin's commands.
+ Choosing a player-specific command will open a second screen, "Player Select", for choosing a player. In this screen you also gain a "next page" and "prev page" buttons for navigating large amounts of players.
+ Choosing the "DeleteAllPlayersData" command will open a warning screen. This will now also appear if you execute this command via chat.

- Removed the command "DeletePlayerData". You shouldn't ever need to delete a specific player's data.

- Added the "_Menu" command for opening the menu.

- Fixed inconsistency - "ResetPlayerTask" used to be executed via "ResetTask". It is now "ResetPlayerTask".

- Added a new "SpecialItemReader" which is responsible for reading items with custom names and lore from the config.

- Fixed: By using the command "RemoveLife", you could spam the player with 0-lives death message by attempting to remove more lives when they're already at 0.

- Added the option to turn off the "has_task" visual effect.

- Updated some default tasks in the config.

+++ src/main/java/org/shonivergames/secretlife/admincommands/DeleteAllPlayersData.java
@@ -3,6 +3,7 @@ package org.shonivergames.secretlife.admincommands;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 import org.shonivergames.secretlife.Main;
+import org.shonivergames.secretlife.PluginMenuManager;
 
 public class DeleteAllPlayersData extends _CommandBase {
     public DeleteAllPlayersData() {
@@ -10,8 +11,12 @@ public class DeleteAllPlayersData extends _CommandBase {
     }
 
     @Override
-    public void executeCommand(CommandSender sender, Player irrelevant) {
-        Main.playerData.deleteAllData();
-        printFeedback(sender);
+    public void executeCommand(CommandSender sender, Player irrelevant, boolean isAfterWarning) {
+        if(isAfterWarning || !(sender instanceof Player)) {
+            Main.playerData.deleteAllData();
+            printFeedback(sender);
+        }
+        else
+            PluginMenuManager.showWarningMenu((Player) sender, this);
     }
 }
